Jqwik Support: Renamed interface for test images
