+++ modules/junit-jqwik/src/test/java/org/testcontainers/junit/jqwik/TestcontainersGroupedRestartedContainerTests.java
@@ -8,7 +8,7 @@ import org.testcontainers.containers.GenericContainer;
 
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.testcontainers.junit.jqwik.JUnitJqwikTestImages.HTTPD_IMAGE;
+import static org.testcontainers.junit.jqwik.JqwikTestImages.HTTPD_IMAGE;
 
 @Disabled("A PropertyLifecyclyeContext does not provide the means to access the parent instance." +
     "This makes it hard to access the parent test instance. The nested container is started and stopped. " +
