+++ modules/junit-jqwik/src/test/java/org/testcontainers/junit/jqwik/TestcontainersRestartBetweenTests.java
@@ -6,7 +6,7 @@ import net.jqwik.api.lifecycle.BeforeProperty;
 import org.testcontainers.containers.GenericContainer;
 
 import static org.junit.jupiter.api.Assertions.*;
-import static org.testcontainers.junit.jqwik.JUnitJqwikTestImages.HTTPD_IMAGE;
+import static org.testcontainers.junit.jqwik.JqwikTestImages.HTTPD_IMAGE;
 
 @Testcontainers
 class TestcontainersRestartBetweenTests {
